requirements.txt
# AI Daily Brief Bot - Python Dependencies
# 安装命令: pip install -r requirements.txt

# 大模型 SDK
zhipuai>=2.1.0

# 语音合成 (TTS)
edge-tts>=6.1.0

# 网页抓取
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
# lxml 在 Windows 上需要编译，如果安装失败可以注释掉
# lxml>=4.9.0

# 定时任务调度
apscheduler>=3.10.0

# 配置管理
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.5.0

# 日志和工具
loguru>=0.7.0

# 额外依赖
sniffio>=1.3.0

